requirements.txt
    alembic==1.17.1
    bidict==0.23.1
    blinker==1.9.0
    click==8.3.0
    dnspython==2.8.0
    eventlet==0.40.3
    Flask==3.1.2
    Flask-Migrate==4.1.0
    Flask-SocketIO==5.5.1
    Flask-SQLAlchemy==3.1.1
    greenlet==3.2.4
    h11==0.16.0
    itsdangerous==2.2.0
    Jinja2==3.1.6
    Mako==1.3.10
    MarkupSafe==3.0.3
    psycopg2-binary==2.9.11
    python-dotenv==1.2.1
    python-engineio==4.12.3
    python-socketio==5.14.3
    simple-websocket==1.1.0
    SQLAlchemy==2.0.44
    typing_extensions==4.15.0
    Werkzeug==3.1.3
    wsproto==1.2.0
    geopy==2.4.1
    haversine==2.8.1
    requests==2.32.3
